add opCodeCache adapter
